Fixed restart command

+++ commands/BOTMANAGER_restart.js
@@ -11,10 +11,12 @@ module.exports = {
 		try{
 		const fs = require('fs');
 		fs.unlinkSync('./runstate.txt')
-		if(safemode = true){
+		if (fs.existsSync(`./safe_mode.flag`)){
 			fs.unlinkSync('./safe_mode.flag')
 		}
+		if (fs.existsSync(`./errorcount.txt`)){
 		fs.unlinkSync('./errorcount.txt')
+		}
 		console.log("A restart has been triggered. Wait 5 seconds before turning it on again.")
 		const { MessageEmbed } = require('discord.js')
 		const RestartedEmbed = new Discord.MessageEmbed()
